(WIP) moving functionality to server - builds
