+++ shared/src/traceData.ts
@@ -25,6 +25,10 @@ export const traceLine = z.object({
       kind: z.number().optional(),
       pos: z.number().optional(),
       end: z.number().optional(),
+      location: z.object({
+        line: z.number(),
+        character: z.number(),
+      }).optional(),
       path: z.string().optional(),
       results: z
         .object({
