+++ src/commands.ts
@@ -4,7 +4,6 @@ import { promisify } from 'node:util'
 import { spawn } from 'node:child_process'
 import { createReadStream, existsSync, readdir as readdirC, statSync } from 'node:fs'
 import * as vscode from 'vscode'
-import { getStatsFromTree, processTraceFiles, processedFiles, showTree, treeIdNodes } from './traceTree'
 import { getTracePanel, prepareWebView } from './webview'
 import { getCurrentConfig } from './configuration'
 import { log } from './logger'
@@ -12,10 +11,9 @@ import type { CommandId } from './constants'
 import { addTraceFile, getWorkspacePath, openTerminal, openTraceDirectoryExternal, setLastMessageTrigger } from './storage'
 import { addTraceDiagnostics, clearTaceDiagnostics } from './traceDiagnostics'
 import { setStatusBarState } from './statusBar'
-import { afterWatches, projectPath, saveName, state, traceFiles, traceRunning } from './appState'
+import { afterWatches, projectPath, saveName, state, traceRunning } from './appState'
 import * as actions from './client/actions'
-
-const readdir = promisify(readdirC)
+import { filterTree } from './client/actions'
 
 const commandHandlers: Record<
   CommandId,
@@ -24,48 +22,9 @@ const commandHandlers: Record<
     'tsperf.tracer.runTrace': () => (...args: unknown[]) => runTrace(args),
     'tsperf.tracer.openInBrowser': (context: vscode.ExtensionContext) => () => prepareWebView(context),
     'tsperf.tracer.gotoTracePosition': (context: vscode.ExtensionContext) => () => gotoTracePosition(context),
-    'tsperf.tracer.sendTrace': () => (event: unknown) => {
-      if (!(Array.isArray(event) && event[0] instanceof vscode.Uri))
-        return
-
-      const fsPath = event[0].fsPath
-
-      sendTrace(dirname(fsPath), basename(fsPath))
-    },
     'tsperf.tracer.openTerminal': () => () => openTerminal(),
     'tsperf.tracer.openTraceDirExternal': () => () => openTraceDirectoryExternal(),
   } as const
-
-async function sendTrace(dirName: string, fileName: string) {
-  const fullFileName = join(dirName, fileName)
-
-  const stream = createReadStream(fullFileName, { autoClose: true, emitClose: true, encoding: 'utf-8' })
-
-  let fileContents = ''
-
-  stream.on('end', () => {
-    addTraceFile(fileName, fileContents)
-    processTraceFiles().then(() => {
-      showTree('check', '', 0)
-
-      clearTaceDiagnostics()
-      for (const editor of vscode.window.visibleTextEditors) {
-        const visibleFileName = editor.document.fileName
-        addTraceDiagnostics(visibleFileName, getStatsFromTree(visibleFileName))
-      }
-    })
-  })
-
-  function readChunks() {
-    const chunk = stream.read()
-    if (chunk === null)
-      return
-
-    fileContents += chunk
-    setImmediate(readChunks)
-  }
-  stream.on('readable', readChunks)
-}
 
 export function registerCommands(context: vscode.ExtensionContext) {
   for (const cmd in commandHandlers) {
@@ -95,13 +54,12 @@ function gotoTracePosition(context: vscode.ExtensionContext) {
   const relativePath = relative(workspacePth, editor.document.fileName)
 
   getTracePanel(context)?.reveal()
-  showTree('', relativePath, startOffset - (editor.document.getText()[startOffset + 1] === '\n' ? 0 : 1))
+  filterTree('', relativePath, startOffset - (editor.document.getText()[startOffset + 1] === '\n' ? 0 : 1), true)
 }
 
 const liveTrace = true // TODO config setting
 async function runTrace(args?: unknown[]) {
   const workspacePath = state.workspacePath.value
-  const { traceCmd } = getCurrentConfig()
 
   let dirName
   if (args && args[0] && Array.isArray(args[0]) && args[0][0] instanceof vscode.Uri) {
@@ -121,10 +79,6 @@ async function runTrace(args?: unknown[]) {
     saveName.value = relative(workspacePath, dirName)
   }
 
-  traceFiles.value = {}
-  processedFiles.clear()
-
-  const newDirName = dirName
   // TODO: use logic from real time metrics that get the tsconfig path
   afterWatches(async () => {
     const traceDir = state.tracePath.value
@@ -140,66 +94,5 @@ async function runTrace(args?: unknown[]) {
       actions.runTrace(workspacePath, traceDir)
       actions.filterTree('checkExpr', '', 0, true)
     }
-    else {
-      const quotedTraceDir = `'${traceDir}'`
-      // eslint-disable-next-line no-template-curly-in-string
-      const fullCmd = `(cd '${newDirName ?? workspacePath}'; ${traceCmd.replace('${traceDir}', quotedTraceDir)})`
-
-      log(fullCmd)
-
-      const newProjectPath = newDirName ?? projectPath.value
-      if (!newProjectPath) {
-        vscode.window.showErrorMessage('could not get project path from workspace folders')
-        return
-      }
-
-      traceRunning.value = true
-
-      traceFiles.value = {}
-      treeIdNodes.clear()
-
-      setStatusBarState('traceError', false)
-
-      log(`shell: ${process.env.SHELL}`)
-      const cmdProcess = spawn(fullCmd, [], { cwd: newProjectPath, shell: process.env.SHELL })
-
-      let err = ''
-      cmdProcess.stderr.on('data', data => err += data.toString())
-
-      cmdProcess.stdout.on('data', data => log(data.toString()))
-
-      cmdProcess.on('error', (error) => {
-        vscode.window.showErrorMessage(error.message)
-      })
-
-      cmdProcess.on('exit', async (code) => {
-        log('---- trace stderr -----')
-        log(err)
-        traceRunning.value = false
-        if (code) {
-          setStatusBarState('traceError', true)
-          vscode.window.showErrorMessage('error running trace')
-          return
-        }
-
-        await sendTraceDir(traceDir)
-      })
-    }
   })
-}
-
-export async function sendTraceDir(traceDir: string) {
-  try {
-    if (!existsSync(traceDir)) {
-      return
-    }
-
-    const fileNames = await readdir(traceDir)
-    for (const fileName of fileNames) {
-      sendTrace(traceDir, fileName)
-    }
-  }
-  catch (e) {
-    vscode.window.showErrorMessage(e instanceof Error ? e.message : `${e}`)
-  }
 }
