+++ src/constants.ts
@@ -25,7 +25,6 @@ export const commandIds = [
   'tsperf.tracer.gotoTracePosition',
   'tsperf.tracer.openInBrowser',
   'tsperf.tracer.runTrace',
-  'tsperf.tracer.sendTrace',
   'tsperf.tracer.openTerminal',
   'tsperf.tracer.openTraceDirExternal',
 ] as const
